requirements.txt
# Core dependencies
psycopg2-binary==2.9.10
aiohttp==3.9.3
psycopg[binary,pool]==3.1.18
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0

# Monitoring and metrics
prometheus-client==0.19.0
prometheus-flask-exporter==0.22.4

# Configuration and validation
jsonschema==4.21.1
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.23.2
pytest-mock==3.12.0
pytest-xdist==3.5.0
pytest-timeout==2.2.0
pytest-env==1.1.3
pytest-randomly==3.15.0
pytest-sugar==0.9.7
pytest-html==4.1.1
pytest-metadata==3.1.0
pytest-benchmark==4.0.0
pytest-clarity==1.0.1
pytest-picked==0.5.0
pytest-watch==4.2.0

# Logging and error handling
structlog==23.2.0
python-json-logger==2.0.7

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# API and web
Flask==3.0.0
gunicorn==21.2.0

# Development tools
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Documentation
Sphinx==7.2.6
sphinx-rtd-theme==1.3.0 